nebula/transcription_docker/requirements.txt
flask
flask-cors
openai
openai-whisper
torch
ffmpeg-python
opencv-python
pillow
python-dotenv
requests
